+++ tests/test_dataloader.py
@@ -5,16 +5,16 @@ import torch
 import pytest
 
 classification_config = EasyDict({
-  'train': {
-    'dataset': 'ImageFolder',
-    'args': {
-      'root': 'tests/test_dataset/classification/train'
+    'train': {
+        'dataset': 'ImageFolder',
+        'args': {
+            'root': 'tests/test_dataset/classification/train'
+        },
     },
-  },
 })
 
 pytorch_loader = EasyDict({
-    'dataloader': 'PytorchDataLoader',
+    'module': 'PytorchDataLoader',
     'args': {
       'num_workers': 1,
       'batch_size': 4,
@@ -23,7 +23,7 @@ pytorch_loader = EasyDict({
   })
 
 dali_loader = EasyDict({
-    'dataloader': 'DALIDataLoader',
+    'module': 'DALIDataLoader',
     'args': {
       'device_id' : 0,
       'num_thread': 1,
@@ -33,15 +33,20 @@ dali_loader = EasyDict({
   })
 
 preprocess_args = EasyDict({
-        'input_size' : 640,
-        'input_normalization' : {
-            'mean' : [0,0,0],
-            'std' : [1, 1, 1],
-            'scaler' : 1
-        },
-    })
+    'input_size' : 640,
+    'input_normalization' : {
+        'mean' : [0,0,0],
+        'std' : [1, 1, 1],
+        'scaler' : 1
+    },
+})
 
-def check_loader(dataloader):
+@pytest.mark.parametrize("dataloader_cfg", [pytorch_loader, dali_loader])
+def test_dataloader(dataloader_cfg):
+    dataloader = create_dataloader(dataloader_config=dataloader_cfg,
+                                   dataset_config=classification_config,
+                                   preprocess_config = preprocess_args,
+                                   collate_fn=None)
     for data in dataloader:
         fetched_data = data
         break
@@ -51,23 +56,3 @@ def check_loader(dataloader):
     assert isinstance(len(dataloader),int)
     assert hasattr(dataloader,'dataset')
     assert isinstance(dataloader.dataset,BasicDatasetWrapper)
-
-@pytest.mark.parametrize("config", [classification_config])
-def test_pytorch_loader(config):
-
-    config.dataloader = pytorch_loader
-    dataloader = create_dataloader(dataset_config=config,
-                                   preprocess_config = preprocess_args,
-                                   collate_fn=None)
-
-    check_loader(dataloader)
-    
-
-@pytest.mark.parametrize("config", [classification_config])
-def test_gpu_dali_loader(config):
-    config.dataloader = dali_loader
-    dataloader = create_dataloader(dataset_config=config,
-                                   preprocess_config = preprocess_args,
-                                   collate_fn=None)
-    check_loader(dataloader)
-