+++ tests/test_nvidia_dali.py
@@ -33,7 +33,7 @@ class_collate_fn = None
 class_dataset_config = EasyDict(
     {
         'train': {
-            'dataset': 'ImageFolder',
+            'name': 'ImageFolder',
             'args': {
                 'root': 'tests/test_dataset/classification/train'
             },
@@ -46,7 +46,7 @@ class_dataset_config.collate_fn = None
 lndmrks_dataset_config = EasyDict(
     {
         'train': {
-            'dataset': 'TestObjDetLandmarksDataset',
+            'name': 'TestObjDetLandmarksDataset',
             'args': {
                 'train': True
             },
@@ -56,7 +56,7 @@ lndmrks_dataset_config = EasyDict(
 lndmrks_dataset_config.collate_fn = 'RetinaFaceCollate'
 
 dali_loader = EasyDict({
-    'dataloader': 'DALIDataLoader',
+    'module': 'DALIDataLoader',
     'args': {
         'device_id' : 0,
         'num_thread': 1,
@@ -66,7 +66,7 @@ dali_loader = EasyDict({
 })
 
 pytorch_loader = EasyDict({
-    'dataloader': 'PytorchDataLoader',
+    'module': 'PytorchDataLoader',
     'args': {
         'batch_size': 4,
         'shuffle': False,
@@ -87,17 +87,18 @@ excpected_raise_error_transform = ['RandomWater','RandomRotate']
 @pytest.mark.parametrize("dataset_config", [class_dataset_config,lndmrks_dataset_config])
 @pytest.mark.parametrize("transform", [transform for transform in transforms])
 def test_dali(dataset_config,transform):
-    dataset_config.dataloader = dali_loader
     augmentations = [EasyDict({'module' : 'nvidia_dali','args' : {'transforms': [transform]}})]
     dataset_config.train.augmentations = augmentations
-    
+
     if dataset_config == lndmrks_dataset_config and transform['transform'] in excpected_raise_error_transform:
         with pytest.raises(RuntimeError):
-            dataloader = create_dataloader(dataset_config=dataset_config,
+            dataloader = create_dataloader(dataloader_config=dali_loader,
+                                        dataset_config=dataset_config,
                                         preprocess_config = preprocess_args,
                                         collate_fn=dataset_config.collate_fn)
     else:
-        dataloader = create_dataloader(dataset_config=dataset_config,
+        dataloader = create_dataloader(dataloader_config=dali_loader,
+                                        dataset_config=dataset_config,
                                         preprocess_config = preprocess_args,
                                         collate_fn=dataset_config.collate_fn)
 
@@ -106,47 +107,43 @@ def test_dali(dataset_config,transform):
             break
     
 def test_neg_dali_aug_and_pytorch_data_loader():
-
-    class_dataset_config.dataloader = pytorch_loader
     augmentations = [EasyDict({'module' : 'nvidia_dali','args' : {'transforms': [transforms[0]]}})]
     class_dataset_config.train.augmentations = augmentations
     
     # Expect RuntimeError if nvidia dali augmentation module used with pytorch dataloader
     with pytest.raises(RuntimeError):
-        dataloader = create_dataloader(dataset_config=class_dataset_config,
+        dataloader = create_dataloader(dataloader_config=pytorch_loader,
+                                        dataset_config=class_dataset_config,
                                         preprocess_config = preprocess_args,
                                         collate_fn=class_dataset_config.collate_fn)
 
 def test_dali_loader_with_no_dali_aug():
-
-    class_dataset_config.dataloader = dali_loader
     augmentations = [EasyDict({'module' : 'albumentations','args' : {'transforms': [transforms[0]]}})]
     class_dataset_config.train.augmentations = augmentations
-    
-    dataloader = create_dataloader(dataset_config=class_dataset_config,
+
+    dataloader = create_dataloader(dataloader_config=dali_loader,
+                                    dataset_config=class_dataset_config,
                                     preprocess_config = preprocess_args,
                                     collate_fn=class_dataset_config.collate_fn)
 
 def test_dali_loader_with_additional_external_aug():
-
-    class_dataset_config.dataloader = dali_loader
     augmentations = [EasyDict({'module' : 'nvidia_dali','args' : {'transforms': [transforms[0]]}}),
                      EasyDict({'module' : 'albumentations','args' : {'transforms': [transforms[0]]}})]
     class_dataset_config.train.augmentations = augmentations
-    
-    dataloader = create_dataloader(dataset_config=class_dataset_config,
+
+    dataloader = create_dataloader(dataloader_config=dali_loader,
+                                    dataset_config=class_dataset_config,
                                     preprocess_config = preprocess_args,
                                     collate_fn=class_dataset_config.collate_fn)
 
 def test_neg_dali_aug_not_the_first_aug():
-
-    class_dataset_config.dataloader = dali_loader
     augmentations = [EasyDict({'module' : 'albumentations','args' : {'transforms': [transforms[0]]}}),
                      EasyDict({'module' : 'nvidia_dali','args' : {'transforms': [transforms[0]]}})]
     class_dataset_config.train.augmentations = augmentations
-    
+
     # Expect RuntimeError if dali augmentations is not in the first order of the augmentation list
     with pytest.raises(RuntimeError):
-        dataloader = create_dataloader(dataset_config=class_dataset_config,
+        dataloader = create_dataloader(dataloader_config=dali_loader,
+                                        dataset_config=class_dataset_config,
                                         preprocess_config = preprocess_args,
                                         collate_fn=class_dataset_config.collate_fn)